Mange APIs by different routes
